Attic unused restraint fxn
